requirements.txt
# ── Core LLM Framework ──
langchain>=0.3.0
langchain-core>=0.3.0
langchain-ollama>=0.3.0

# ── Math Safety ──
asteval

# ── Environment ──
python-dotenv

# ── Voice (STT & TTS) — Phase 3 ──
# faster-whisper
# piper-tts
# sounddevice
# numpy

# ── Testing ──
pytest
pytest-asyncio
